Upgrade cocos2d to 2.0 API + types.
[UPDATED]   Cocos2D code now uses the 2.0 API and OpenGL/ES 2.0.
[IMPROVED]  Better type handling through more type warning fixes.
[REMOVED]   PearlConfig.visualFx - a user shouldn't have to worry about
            tweaking config for decent performance.

+++ Pearl-Cocos2D/View/PearlCCBarLayer.h
@@ -31,7 +31,7 @@
     CCMenu                 *_menuMenu;
     CCLabelTTF             *_messageLabel;
 
-    long                  _textColor, _renderColor;
+    NSUInteger            _textColor, _renderColor;
     CGPoint               _showPosition;
 
     BOOL                 _dismissed;
@@ -39,8 +39,8 @@
 
 @property (nonatomic, readonly) BOOL dismissed;
 
-+ (PearlCCBarLayer *)barWithColor:(long)aColor position:(CGPoint)aShowPosition;
--(id) initWithColor:(long)aColor position:(CGPoint)aShowPosition;
++ (PearlCCBarLayer *)barWithColor:(NSUInteger)aColor position:(CGPoint)aShowPosition;
+-(id) initWithColor:(NSUInteger)aColor position:(CGPoint)aShowPosition;
 
 -(void) setButtonTitle:(NSString *)aTitle callback:(id)target :(SEL)selector;
 -(CGPoint) hidePosition;
